+++ src/zabbix_enums/common/alert.py
