+++ src/zabbix_enums/common/discovered_host.py
