+++ src/zabbix_enums/common/discovered_service.py
