+++ src/zabbix_enums/common/discovery_rule.py
@@ -1,3 +1,3 @@
-from ._common import EntityStatus
+from .._common import EntityStatus
 
 DiscoveryRuleStatus = EntityStatus
