+++ src/zabbix_enums/common/host_group.py
@@ -1,5 +1,5 @@
 from enum import IntEnum
-from ._common import DiscoveryFlag, NoYes
+from .._common import DiscoveryFlag, NoYes
 
 
 HostGroupFlag = DiscoveryFlag
