+++ src/zabbix_enums/common/host_prototype.py
@@ -1,6 +1,6 @@
 from enum import IntEnum
-from ._common_host import HostStatus, HostInventoryMode
-from ._common import PrototypeDiscover
+from .host import HostStatus, HostInventoryMode
+from .._common import PrototypeDiscover
 
 HostPrototypeStatus = HostStatus
 HostPrototypeInventoryMode = HostInventoryMode
