+++ src/zabbix_enums/common/image.py
