+++ src/zabbix_enums/common/item_http.py
@@ -1,5 +1,5 @@
 from enum import IntEnum
-from ._common import NoYes
+from .._common import NoYes
 
 
 ItemAllowTraps = NoYes
