+++ src/zabbix_enums/common/script.py
