+++ src/zabbix_enums/common/snmp.py
