+++ src/zabbix_enums/common/task.py
